Split pytorch_ops into a handler package
Convert execution/pytorch_ops.py (~1.9k lines) into a package with focused
modules: _registry (handler dispatch + execute_graph), _common (shared
tensor/graph helpers), _warnings (reference-path warnings), and per-op
handler modules under handlers/ (conv, matmul, batchnorm, norm, reduction,
resample, sdpa, pointwise). Handlers self-register via the package __init__;
public API and behavior are unchanged.
